:art: Improve Owasm ext code structure and fix partial cmp bug
