+++ owasm/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "owasm"
-version = "0.1.8"
+version = "0.1.9"
 authors = ["Band Protocol <connect@bandprotocol.com>"]
 edition = "2018"
 license = "Apache-2.0"
